move `lints()` to `iter_count.rs`
